+++ examples/example-antd-management-fast-framework/src/businessComponents/SelectField/SelectDrawerField/index.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import { connect } from 'easy-soft-dva';
-
 import { FieldExtra } from 'antd-management-fast-component';
 
 import { SimpleMultiPageSingleSelectDrawer } from '../../Drawers/Multi/SelectDrawer/SimpleMultiPageSingleSelectDrawer';
@@ -10,10 +8,6 @@ const {
   SelectFieldExtra: { BaseSelectFieldExtra },
 } = FieldExtra;
 
-@connect(({ smsCategory, schedulingControl }) => ({
-  smsCategory,
-  schedulingControl,
-}))
 class SelectDrawerField extends BaseSelectFieldExtra {
   selectValueText = (data) => {
     const { title } = {