+++ examples/example-antd-management-fast-framework/src/businessComponents/SelectField/SelectModalField/index.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import { connect } from 'easy-soft-dva';
-
 import { FieldExtra } from 'antd-management-fast-component';
 
 import { SimpleSingleSelectModal } from '../../Modals/SimpleSingleSelectModal';
@@ -10,10 +8,6 @@ const {
   SelectFieldExtra: { BaseSelectFieldExtra },
 } = FieldExtra;
 
-@connect(({ smsCategory, schedulingControl }) => ({
-  smsCategory,
-  schedulingControl,
-}))
 class SelectModalField extends BaseSelectFieldExtra {
   selectValueText = (data) => {
     const { title } = {