+++ packages/antd-management-fast-charts/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 2.6.174
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.7.67
+
+## 2.6.173
+
+### Patch Changes
+
+- Updated dependencies
+  - antd-management-fast-component@2.7.66
+
 ## 2.6.172
 
 ### Patch Changes