+++ packages/antd-management-fast-component/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 2.7.67
+
+### Patch Changes
+
+- update
+
+## 2.7.66
+
+### Patch Changes
+
+- update
+
 ## 2.7.65
 
 ### Patch Changes